Fix failing tests
Fix tests
