+++ node-graph/nodes/repeat/src/repeat_nodes.rs
@@ -43,7 +43,7 @@ async fn repeat<T: Into<Graphic> + Default + Send + Clone + 'static>(
 }
 
 #[node_macro::node(category("Repeat"))]
-async fn repeat_array<T: Into<Graphic> + Default + Send + Clone + 'static>(
+pub async fn repeat_array<T: Into<Graphic> + Default + Send + Clone + 'static>(
 	ctx: impl ExtractAll + CloneVarArgs + Ctx,
 	#[implementations(
 		Context -> Table<Graphic>,
@@ -177,16 +177,23 @@ async fn repeat_on_points<T: Into<Graphic> + Default + Send + Clone + 'static>(
 #[cfg(test)]
 mod test {
 	use super::*;
-	use crate::generator_nodes::RectangleNode;
 	use core_types::Ctx;
 	use core_types::Node;
+	use core_types::transform::Footprint;
 	use glam::DVec2;
 	use graphene_core::ReadPositionNode;
 	use graphene_core::extract_xy::{ExtractXyNode, XY};
 	use graphic_types::Vector;
+	use kurbo::Shape;
+	use kurbo::{BezPath, DEFAULT_ACCURACY, Rect};
 	use std::future::Future;
 	use std::pin::Pin;
+	use vector_nodes::generator_nodes::RectangleNode;
 	use vector_types::subpath::Subpath;
+
+	fn vector_node_from_bezpath(bezpath: BezPath) -> Table<Vector> {
+		Table::new_from_element(Vector::from_bezpath(bezpath))
+	}
 
 	#[derive(Clone)]
 	pub struct FutureWrapperNode<T: Clone>(T);
@@ -200,8 +207,8 @@ mod test {
 	}
 
 	#[tokio::test]
-	async fn instance_on_points_test() {
-		let owned = OwnedContextImpl::default().into_context();
+	async fn repeat_on_points_test() {
+		let context = OwnedContextImpl::default().into_context();
 		let rect = RectangleNode::new(
 			FutureWrapperNode(()),
 			ExtractXyNode::new(ReadPositionNode::new(FutureWrapperNode(()), FutureWrapperNode(0)), FutureWrapperNode(XY::Y)),
@@ -213,12 +220,72 @@ mod test {
 
 		let positions = [DVec2::new(40., 20.), DVec2::ONE, DVec2::new(-42., 9.), DVec2::new(10., 345.)];
 		let points = Table::new_from_element(Vector::from_subpath(Subpath::from_anchors(positions, false)));
-		let generated = super::instance_on_points(owned, points, &rect, false).await;
+		let generated = super::repeat_on_points(context, points, &rect, false).await;
 		assert_eq!(generated.len(), positions.len());
 		for (position, generated_row) in positions.into_iter().zip(generated.iter()) {
 			let bounds = generated_row.element.bounding_box_with_transform(*generated_row.transform).unwrap();
 			assert!(position.abs_diff_eq((bounds[0] + bounds[1]) / 2., 1e-10));
 			assert_eq!((bounds[1] - bounds[0]).x, position.y);
+		}
+	}
+
+	#[tokio::test]
+	async fn repeat() {
+		let direction = DVec2::X * 1.5;
+		let count = 3;
+		let context = OwnedContextImpl::default().into_context();
+		let repeated = super::repeat_array(
+			context,
+			&FutureWrapperNode(vector_node_from_bezpath(Rect::new(0., 0., 1., 1.).to_path(DEFAULT_ACCURACY))),
+			direction,
+			0.,
+			count,
+		)
+		.await;
+		let vector_table = vector_nodes::flatten_path(Footprint::default(), repeated).await;
+		let vector = vector_table.iter().next().unwrap().element;
+		assert_eq!(vector.region_manipulator_groups().count(), 3);
+		for (index, (_, manipulator_groups)) in vector.region_manipulator_groups().enumerate() {
+			assert!((manipulator_groups[0].anchor - direction * index as f64 / (count - 1) as f64).length() < 1e-5);
+		}
+	}
+
+	#[tokio::test]
+	async fn repeat_transform_position() {
+		let direction = DVec2::new(12., 10.);
+		let count = 8;
+		let context = OwnedContextImpl::default().into_context();
+		let repeated = super::repeat_array(
+			context,
+			&FutureWrapperNode(vector_node_from_bezpath(Rect::new(0., 0., 1., 1.).to_path(DEFAULT_ACCURACY))),
+			direction,
+			0.,
+			count,
+		)
+		.await;
+		let vector_table = vector_nodes::flatten_path(Footprint::default(), repeated).await;
+		let vector = vector_table.iter().next().unwrap().element;
+		assert_eq!(vector.region_manipulator_groups().count(), 8);
+		for (index, (_, manipulator_groups)) in vector.region_manipulator_groups().enumerate() {
+			assert!((manipulator_groups[0].anchor - direction * index as f64 / (count - 1) as f64).length() < 1e-5);
+		}
+	}
+
+	#[tokio::test]
+	async fn repeat_radial() {
+		let context = OwnedContextImpl::default().into_context();
+		let repeated = super::repeat_radial(context, &FutureWrapperNode(vector_node_from_bezpath(Rect::new(-1., -1., 1., 1.).to_path(DEFAULT_ACCURACY))), 45., 4., 8).await;
+		let vector_table = vector_nodes::flatten_path(Footprint::default(), repeated).await;
+		let vector = vector_table.iter().next().unwrap().element;
+		assert_eq!(vector.region_manipulator_groups().count(), 8);
+
+		for (index, (_, manipulator_groups)) in vector.region_manipulator_groups().enumerate() {
+			let expected_angle = (index as f64 + 1.) * 45.;
+
+			let center = (manipulator_groups[0].anchor + manipulator_groups[2].anchor) / 2.;
+			let actual_angle = DVec2::Y.angle_to(center).to_degrees();
+
+			assert!((actual_angle - expected_angle).abs() % 360. < 1e-5, "Expected {expected_angle} found {actual_angle}");
 		}
 	}
 }
